(Ugly) first pass at new event parser.
Parsing events pages on the 2012 SXSW web site is a much uglier
business than it was for the 2011 version of the site. Almost all of
the nice <... class="xyz"> attributes from the 2011 version of the
HTML are gone; they've been replaced with generic <div> and <span>
elements whose contents must be gleaned by looking for special text
values, or at certain fixed positions in the element stream. (I hope
they're fixed positions! I've only tested on one or 2 example pages so
far.)

What I'm checking in here is a very ugly, but mostly functional,
initial pass at a new parser for the 2012 event pages. There are some
unused leftover bits from the 2011 parser that I'll clean up later.
There's probably a cleaner way to implement many of the new parsing
functions. And the nice date+time parser from the 2011 version won't
work with the new date and time strings, so it needs to be rewritten.
I'm checking it in as-is for now, along with a BUG notice that it's
not working with the 2012 site.

I also added a stupid "debugparse" command that simply returns String
values, so that I can use "show" to do a bit of printf-style debugging
of various TagSoup parsing incantations as I go. The DebugParse module
is ugly and should either be rewritten or thrown out when I'm finished
with the production parser.