Add a batchparse command.
This command can parse one or more event HTML files from the web site,
or from local files (e.g., created with the dump or batchdump
command).

Also, changed the uriToFileName function so that it can handle http://
URLs or file paths: it just strips off the file part of the (HTTP or
local filesystem) path.