Not all events have a description.
The new description parser is slightly more robust: in addition to
catching events with no description, it's also smarter about searching
just the "data clear" div for the description text. It does this by
looking for the "<!-- eo data -->" comment marking the ending </div>
for this section. (This heuristic wouldn't be necessary if I were
using a real parser, but TagSoup doesn't retain hierarchical
document structure.)