Some events don't have a start time.
Don't have any good ideas about how to deal with these events. The
best I can do is to specify on which day they occur. For now, treat
these fields as Maybe UTCTime. Note that this produces
"Just":"somedate" fields in the JSON output when the date is
specified. Not sure yet whether to fix this in the JSON output, or in
the Event type.