Improve JSON encoding for Event data type.
Specifically, Maybe a types now encode to the unadorned value when
wrapped by a Just, and a JSON null value when they are
Nothing. (Previously, they would encode to {"Just": "foo"} when
wrapped by Just, and "Nothing" otherwise.)

Note that simply adding a "deriving Generic" clause to Event and
declaring a ToJSON instance was sufficient to realize this change. I
found this surprising, as it wasn't clear from the Aeson documentation
that making only that change would produce a different result than the
Data.Aeson.Generic toJSON function that I was using before this
change. (I had expected to have to write a custom toJSON method for
the ToJSON Event instance.)