Add sxred, a redis import client.
All it does right now is import all events URLs into a Redis store.
This is mainly just a proof-of-concept that the Redis.hs code is still
working. Note that the Haskell redis package has been deprecated in
favor of hedis, so the next step is probably going to be to port the
existing basic Redis functionality to the new library.