Don't import artist description from event pages into Redis.
It's usually not very informative and/or useful, and consumes a lot of
space. Removing it should shorten client downloads significantly. (It
reduces the in-memory footprint of the entire 2012 event database in
the Redis server from about 48MB to 3.4MB.) The client can still see
the description by visiting the official SXSW event page, anyway.