Update for 2013 SXSW music showcase schedule.
Note that the schedule is still under construction at the time of this
check-in: there are many events without an exact start time, and about
a handful with no venue. The parser can handle the former, but not the
latter, as the data model for the iOS client assumes that each event
has a venue. Note that the sxcrape "parse" and "batchparse" commands
will fail hard on events with no venues. Presumably these events will
be fixed by the time the final schedule is ready; for now, I just
ignore them.